[IMP] 优化单元测试代码

+++ mybatisgx-core/src/test/resources/db/migration/pgsql/V1.0.0__init_simple.sql
@@ -1,6 +1,6 @@
 create table test_user_simple
 (
-    id             NUMBER(19)   not null,
+    id             bigint       not null,
     role_ids       varchar(256) not null,
     name           varchar(256) null,
     age            int          null,
@@ -9,9 +9,9 @@ create table test_user_simple
     user_name      varchar(256) not null,
     password       varchar(256) not null,
 
-    input_user_id  NUMBER(19)   not null,
+    input_user_id  bigint       not null,
     input_time     TIMESTAMP    not null,
-    update_user_id NUMBER(19)   null,
+    update_user_id bigint       null,
     update_time    TIMESTAMP    null,
     status         int          not null,
     version        int          not null,
@@ -20,34 +20,34 @@ create table test_user_simple
 
 create table user_detail_simple
 (
-    id             NUMBER(19)  not null,
+    id             bigint      not null,
     code           varchar(64) not null,
-    user_id        NUMBER(19)  not null,
+    user_id        bigint      not null,
 
-    input_user_id  NUMBER(19)  not null,
+    input_user_id  bigint      not null,
     input_time     TIMESTAMP   not null,
-    update_user_id NUMBER(19)  null,
+    update_user_id bigint      null,
     update_time    TIMESTAMP   null,
     primary key (id)
 );
 
 create table user_role_simple
 (
-    id      NUMBER(19) not null,
-    user_id NUMBER(19) not null,
-    role_id NUMBER(19) null,
+    id      bigint not null,
+    user_id bigint not null,
+    role_id bigint null,
     primary key (id)
 );
 
 create table role_simple
 (
-    id             NUMBER(19)  not null,
+    id             bigint      not null,
     name           varchar(64) not null,
     code           varchar(64) null,
 
-    input_user_id  NUMBER(19)  not null,
+    input_user_id  bigint      not null,
     input_time     TIMESTAMP   not null,
-    update_user_id NUMBER(19)  null,
+    update_user_id bigint      null,
     update_time    TIMESTAMP   null,
     status         int         not null,
     version        int         not null,
